Address Copilot review feedback
